Organize project scturcure
