--- pdf_merger.pyw
@@ -1,5 +1,0 @@
-from gui import Application
-
-app = Application()
-app.master.title("PDF Merger")
-app.mainloop()
