+++ simple_pdf_merger/pdf_merger.pyw
@@ -1,0 +1,5 @@
+from application_gui import ApplicationGUI
+
+app = ApplicationGUI()
+app.master.title("PDF Merger")
+app.mainloop()
